VOICE-2: the .tlk speech file was indexed at the wrong stride too
Fixing VOICE-1 let the voice system get further, and it immediately hit the next
instance of the same defect -- six new assertions appeared in voicemanager.cpp
(1787..1816) that had never been reachable before.

TlkFile reads a 32-bit Windows file with sizeof(long) widths, and gets it wrong
in three separate places:

  Index2Data()   strided the index table by sizeof(long) = 8, where the file
                 stores 4-byte entries, so every entry after the first was read
                 at the wrong offset;
  GetFragIndex() then read 8 bytes over a 4-byte field, returning two
                 concatenated table entries as one huge index -- which is why
                 the following GetData() returned NULL and tripped the asserts;
  TlkBlock       declared filelen/compressedlen as unsigned long, putting the
                 audio `data` at offset 16 instead of 8.

All three now use explicit 32-bit types.

Assertion inventory for an identical campaign flight, before and after this pair
of fixes (VOICE-1 + VOICE-2):

  before   voicefilter.cpp:992, :1095          4 lines
  after    (gone)
  before   voicemanager.cpp:1787..1816         12 lines (revealed by VOICE-1)
  after    (gone)

Remaining recurring assertions in that run, for the next pass: cmpclass.cpp:2003
(campaign save encode size mismatch -- SAVE-1 family, and the most interesting of
the three), mission.cpp:2040, atm.cpp:2134, render2d.cpp:1163.

+++ src/falcsnd/voicemanager.h
@@ -182,20 +182,28 @@ public:
 
 class TlkFile : public FileMemMap
 {
+    /* FF_LINUX: .tlk is a 32-bit Windows file, so every width here is 4 bytes,
+     * not sizeof(long). Left as long this went wrong three times over:
+     *   - Index2Data strided the index table by 8 instead of 4, so every entry
+     *     after the first was read at the wrong offset;
+     *   - GetFragIndex then read 8 bytes over a 4-byte field, returning two
+     *     concatenated entries as one huge index;
+     *   - TlkBlock put `data` at offset 16 instead of 8.
+     * Same defect as VOICE-1 in voicefilter.h, one file over. */
     long Index2Data(int tlkind)
     {
-        return TLK_HEADER_INFO + sizeof(long) * tlkind;
+        return TLK_HEADER_INFO + sizeof(int32_t) * tlkind;
     };
     long GetFragIndex(int tlkind)
     {
-        BYTE *data = GetData(Index2Data(tlkind), sizeof(long));
+        BYTE *data = GetData(Index2Data(tlkind), sizeof(int32_t));
         ShiAssert(data not_eq NULL);
-        return data ? *(long *)data : 0;
+        return data ? *(int32_t *)data : 0;
     };
     struct TlkBlock
     {
-        unsigned long filelen;
-        unsigned long compressedlen;
+        uint32_t filelen;
+        uint32_t compressedlen;
         char data[1]; // more in practice
     };
 public:
